+++ libifstate/util.py
@@ -10,11 +10,13 @@ from pyroute2.netlink import NLM_F_EXCL
 
 
 class IPRouteExt(IPRoute):
-    def del_filter_by_info(self, index=0, handle=0, info=0):
+    def del_filter_by_info(self, index=0, handle=0, info=0, parent=0):
         msg = tcmsg()
         msg['index'] = index
         msg['handle'] = handle
         msg['info'] = info
+        if parent != 0:
+            msg['parent'] = parent
 
         return tuple(ipr.nlm_request(
             msg,